fix(Navbar): add conditional rendering for navigation links and login button based on user authentication status
The changes in this commit add conditional rendering to the Navbar component in order to display navigation links and a login button based on the user's authentication status.

Previously, all navigation links were always rendered, regardless of whether the user was authenticated or not. With this commit, the navigation links will only be rendered if the user is authenticated. This improves the user experience by showing relevant links based on the user's authentication status.

Additionally, a login button is now conditionally rendered based on whether the user is authenticated or not. If the user is not authenticated, the login button will be displayed, allowing them to log in. If the user is already authenticated, the login button will not be shown.

These changes were made to provide a more personalized and intuitive navigation experience for users of the application.